pyproject.toml
[project]
name = "ttclock"
version = "3.0.4"
description = "Time Tracking Clockin Automations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "attrs==25.1.0",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "h11==0.14.0",
    "idna==3.10",
    "outcome==1.3.0.post0",
    "packaging==24.2",
    "playwright==1.48.0",
    "pysocks==1.7.1",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "typing-extensions==4.12.2",
    "urllib3==2.3.0",
]

[project.scripts]
ttclock = "ttclock.__main__:main"

[tool.hatch.build]
include = [
    "ttclock/",
    ".ittclock.env.example",
    "README.md",
]
exclude = [
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
